+++ docs/changelog.md
@@ -4,6 +4,9 @@
 * Added a linter
 * Made changes based on the linters output
 
+### Features
+* API container now saves free IPs on a local database
+
 ### Fixes
 * Fix go.mod for commons & cli to reflect monorepo and replaced imports with write package name